+++ R/count_linkage.R
@@ -2,7 +2,7 @@
 #' 
 #' Count occurrences of linkages across multiple domino results from a linkage summary
 #' 
-#' @param linkage_summary a [linkage_summary()] object
+#' @param link_summary a [linkage_summary()] object
 #' @param cluster the name of the cell cluster being compared across multiple domino results
 #' @param group.by the name of the column in `linkage_summary@subject_meta` by which to group subjects for counting.
 #'  If NULL, only total counts of linkages for linkages in the cluster across all subjects is given.
@@ -18,11 +18,11 @@
 #' @examples
 #' data(LinkageSummary)
 #' count_linkage(
-#'   linkage_summary = LinkageSummary$linkage_sum_tiny, cluster = "C1", 
+#'   link_summary = LinkageSummary$linkage_sum_tiny, cluster = "C1", 
 #'   group.by = "group", linkage = "rec")
 #' 
-count_linkage <- function(linkage_summary, cluster, group.by = NULL, linkage = "rec_lig", subject_names = NULL) {
-    check_arg(linkage_summary, allow_class = "linkage_summary", allow_len = 1)
+count_linkage <- function(link_summary, cluster, group.by = NULL, linkage = "rec_lig", subject_names = NULL) {
+    check_arg(link_summary, allow_class = "linkage_summary", allow_len = 1)
     check_arg(cluster, allow_class = "character", allow_len = 1)
     check_arg(group.by, allow_class = c("character", "NULL"), allow_len = c(0, 1))
     check_arg(linkage, allow_class = "character", allow_len = 1,
@@ -30,26 +30,26 @@ count_linkage <- function(linkage_summary, cluster, group.by = NULL, linkage = "
     check_arg(subject_names, allow_class = c("factor", "character", "NULL"))
 
     if (is.null(subject_names)) {
-        subject_names <- linkage_summary@subject_names
+        subject_names <- link_summary@subject_names
     }
-    all_int_ls <- lapply(linkage_summary@subject_linkages, FUN = function(x) {
+    all_int_ls <- lapply(link_summary@subject_linkages, FUN = function(x) {
         return(x[[cluster]][[linkage]])
     })
     all_int <- unlist(all_int_ls)
     feature <- table(unlist(all_int))
     dframe <- data.frame(feature = names(feature), total_count = as.numeric(feature))
     if (!is.null(group.by)) {
-        if (!group.by %in% colnames(linkage_summary@subject_meta)) {
+        if (!group.by %in% colnames(link_summary@subject_meta)) {
             stop("group.by variable not present in subject_meta")
         }
-        groups <- levels(factor(linkage_summary@subject_meta[[group.by]]))
+        groups <- levels(factor(link_summary@subject_meta[[group.by]]))
         for (g in groups) {
-            g_index <- linkage_summary@subject_meta[[group.by]] == g
-            g_subjects <- linkage_summary@subject_meta[g_index, 1]
+            g_index <- link_summary@subject_meta[[group.by]] == g
+            g_subjects <- link_summary@subject_meta[g_index, 1]
             int_count <- list()
             for (f in dframe[["feature"]]) {
                 count <- vapply(g_subjects, FUN.VALUE = logical(1), FUN = function(x) {
-                    f %in% linkage_summary@subject_linkages[[x]][[cluster]][[linkage]]
+                    f %in% link_summary@subject_linkages[[x]][[cluster]][[linkage]]
                 })
                 int_count[[f]] <- sum(count)
             }